third_party/WebKit/LayoutTests/inspector/elements/styles-2/cssom-shorthand-important.html

<html>
<head>
<style>
#inspected {
  padding: 10px 50px !important;
}
</style>

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

document.styleSheets[0].cssRules[0].style.marginTop = "10px"

function test()
{
    InspectorTest.selectNodeAndWaitForStyles("inspected", dump);

    function dump()
    {
        InspectorTest.dumpSelectedElementStyles(true);
        InspectorTest.completeTest();
    }
}
</script>
</head>

<body onload="runTest()">
<p>Tests that CSSOM-modified shorthands are reporting their "important" bits.</p>
<div id="inspected">Text</div>
</body>
</html>
