third_party/WebKit/LayoutTests/inspector/elements/styles-2/cssom-shorthand-important-expected.txt

Tests that CSSOM-modified shorthands are reporting their "important" bits.

Text
[expanded] 
element.style { ()

[expanded] 
#inspected { (<style>…</style>)
    margin-top: 10px;
    padding: 10px 50px !important;
        padding-top: 10px;
        padding-right: 50px;
        padding-bottom: 10px;
        padding-left: 50px;

[expanded] 
div { (user agent stylesheet)
    display: block;
